Extract ChatMessage to a separate file
