+++ frontend/src/Chat.tsx
@@ -1,38 +1,6 @@
 import * as React from 'react';
-import classNames from 'classnames';
-import {IGameMode} from './types';
+import ChatMessage, {Message} from './ChatMessage';
 
-export type MessageVariant =
-  | 'ResettingGame'
-  | 'StartingGame'
-  | {TrickWon: {winner: number; points: number}}
-  | {RankAdvanced: {player: number; new_rank: number}}
-  | {NewLandlordForNextGame: {landlord: number}}
-  | {PointsInKitty: {points: number; multiplier: number}}
-  | {JoinedGame: {player: number}}
-  | {JoinedTeam: {player: number}}
-  | {LeftGame: {name: string}}
-  | {KittySizeSet: {size: number | null}}
-  | {NumDecksSet: {num_decks: number | null}}
-  | {NumFriendsSet: {num_friends: number | null}}
-  | {GameModeSet: {game_mode: IGameMode}}
-  | 'TookBackPlay'
-  | {SetDefendingPointVisibility: {visible: boolean}}
-  | {SetLandlord: {landlord: number | null}}
-  | {SetRank: {rank: string}}
-  | {MadeBid: {card: string; count: number}};
-
-export type BroadcastMessage = {
-  actor: number;
-  variant: MessageVariant;
-};
-
-export type Message = {
-  from: string;
-  message: string;
-  from_game?: boolean;
-  data?: BroadcastMessage;
-};
 type Props = {
   messages: Message[];
 };
@@ -69,12 +37,7 @@ const Chat = (props: Props) => {
     <div className="chat">
       <div className="messages">
         {props.messages.map((m, idx) => (
-          <p
-            key={idx}
-            className={classNames('message', {'game-message': m.from_game})}
-          >
-            {m.from}: {m.message}
-          </p>
+          <ChatMessage message={m} key={idx} />
         ))}
         <div className="chat-anchor" ref={anchor} />
       </div>
